+++ pocket/categories/cmdi.py
@@ -1,8 +1,8 @@
 from cmd2 import Cmd
 
+
 class cmdi(Cmd):
-
-
+    pass
 
 
 def cmdiExec():
