+++ pocket/categories/sqli.py
@@ -1,0 +1,9 @@
+from cmd2 import Cmd
+
+
+class sqli(Cmd):
+    pass
+
+
+def sqliExec():
+    sqli().cmdloop()
