+++ pocket/utilities/frogs.py
@@ -1,0 +1,52 @@
+import random
+
+
+def getFrog():
+    frogs = {
+        0: """
+         @..@         /===================\\
+        (\\--/)        |     Welcome to    |
+       (.>__<.)       |       Pocket      |
+       ^^^  ^^^       \\===================/
+   \n""",
+        1: """      
+         @..@         /===================\\
+        (____)        |     Welcome to    |
+       (>____<)       |       Pocket      |
+        ^^~~^^        \\===================/
+    \n""",
+        2: """
+         00           /===================\\
+        (--)          |     Welcome to    |
+       ( || )         |       Pocket      |
+       ^^~~^^         \\===================/
+   \n""",
+        3: """
+       (0) (0)        /===================\\
+      ( ----- )       |     Welcome to    |
+       [ | | ]        |       Pocket      |
+      ~^~   ~^~       \\===================/
+    \n""",
+        4: """
+       \\@--@/         /===================\\
+      [ .~~. ]        |     Welcome to    |
+       ||  ||         |       Pocket      |
+      ~~~  ~~~        \\===================/
+
+    \n""",
+        5: """
+        Welcome To
+ ██▓███   ▒█████   ▄████▄   ██ ▄█▀▓█████▄▄▄█████▓
+▓██░  ██▒▒██▒  ██▒▒██▀ ▀█   ██▄█▒ ▓█   ▀▓  ██▒ ▓▒
+▓██░ ██▓▒▒██░  ██▒▒▓█    ▄ ▓███▄░ ▒███  ▒ ▓██░ ▒░
+▒██▄█▓▒ ▒▒██   ██░▒▓▓▄ ▄██▒▓██ █▄ ▒▓█  ▄░ ▓██▓ ░ 
+▒██▒ ░  ░░ ████▓▒░▒ ▓███▀ ░▒██▒ █▄░▒████▒ ▒██▒ ░ 
+▒▓▒░ ░  ░░ ▒░▒░▒░ ░ ░▒ ▒  ░▒ ▒▒ ▓▒░░ ▒░ ░ ▒ ░░   
+░▒ ░       ░ ▒ ▒░   ░  ▒   ░ ░▒ ▒░ ░ ░  ░   ░    
+░░       ░ ░ ░ ▒  ░        ░ ░░ ░    ░    ░      
+             ░ ░  ░ ░      ░  ░      ░  ░        
+               ░                              
+        """
+    }
+
+    return frogs[random.randint(0, len(frogs))]
