--- src/categories/sqli.py
@@ -1,9 +1,0 @@
-from cmd2 import Cmd
-
-class sqli(Cmd):
-
-
-
-
-def sqliExec():
-    sqli().cmdloop()
